+++ tests/cython/test_protocol.py
@@ -5,10 +5,9 @@ import pytest
 import uvloop
 
 import stario.responses as responses
-from stario import App
 from stario.http.compression import CompressionConfig
 from stario.telemetry.noop import NoOpTracer
-
+from stario_cython.app import App
 from stario_cython.protocol import HttpProtocol